Fixed bullets so that they would play well with { }.
We can now have script like
 assert P.
  { destruct n.
     - solve_case1.
     - solve_case2.
  }
  solve_goal

However there is an undesirable interaction with Focus (which we might, anyway, consider deprecated in favour of {}). Indeed, for compatibility with v8.3, Unfocus is called implicitely after each proof command if there is no focused goal. And the new behaviour of bullets is to allow arbitrary unfocusing command "pass trough" them. As a result, a script like
Focus.
  split
  - solves_first_goal
will result in a fully unfocused proof.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@14262 85f007b7-540e-0410-9357-904b9bb8a0f7