Passed conv_algo to evar_define and move call to solve_refl to evar_define so that it can recursively deal with evar/evar problems. Also, check_evar_instance now called after each instantiation.
Also did a bit of file reformatting.

The commit apparently induces a loss of some 0,4% on the compilation
of the standard library. Maybe, introducing a heuristic to decide when
to call check_evar_instance (which I guess is responsible for the
overhead) might be a good thing to look at?

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@14534 85f007b7-540e-0410-9357-904b9bb8a0f7